# Speed up writing into `io::Write` e.g. `Vec<u8>`