--- a/testing/benches/all.rs
+++ b/testing/benches/all.rs
@@ -1,28 +1,39 @@
-#[macro_use]
-extern crate criterion;
+use std::hint::black_box;
+use std::iter::repeat;
 
-use criterion::Criterion;
+use criterion::{criterion_group, criterion_main, Criterion};
 use rinja::Template;
 
 criterion_main!(benches);
 criterion_group!(benches, functions);
 
 fn functions(c: &mut Criterion) {
-    c.bench_function("Big table", |b| big_table(b, 100));
+    c.bench_function("Big table", big_table);
+    c.bench_function("Big table (fmt)", big_table_fmt);
+    c.bench_function("Big table (io)", big_table_io);
+
     c.bench_function("Teams", teams);
+    c.bench_function("Teams (fmt)", teams_fmt);
+    c.bench_function("Teams (io)", teams_io);
 }
 
-fn big_table(b: &mut criterion::Bencher, size: usize) {
-    let mut table = Vec::with_capacity(size);
-    for _ in 0..size {
-        let mut inner = Vec::with_capacity(size);
-        for i in 0..size {
-            inner.push(i);
-        }
-        table.push(inner);
-    }
-    let ctx = BigTable { table };
-    b.iter(|| ctx.render().unwrap());
+fn big_table(b: &mut criterion::Bencher) {
+    let ctx = BigTable::default();
+    b.iter(|| black_box(&ctx).render().unwrap());
+}
+
+fn big_table_fmt(b: &mut criterion::Bencher) {
+    let ctx = BigTable::default();
+    b.iter(|| black_box(&ctx).to_string());
+}
+
+fn big_table_io(b: &mut criterion::Bencher) {
+    let ctx = BigTable::default();
+    b.iter(|| {
+        let mut vec = Vec::with_capacity(BigTable::SIZE_HINT);
+        black_box(&ctx).write_into(&mut vec).unwrap();
+        vec
+    });
 }
 
 #[derive(Template)]
@@ -31,29 +42,33 @@ struct BigTable {
     table: Vec<Vec<usize>>,
 }
 
+impl Default for BigTable {
+    fn default() -> Self {
+        const SIZE: usize = 100;
+
+        BigTable {
+            table: repeat((0..SIZE).collect()).take(SIZE).collect(),
+        }
+    }
+}
+
 fn teams(b: &mut criterion::Bencher) {
-    let teams = Teams {
-        year: 2015,
-        teams: vec![
-            Team {
-                name: "Jiangsu".into(),
-                score: 43,
-            },
-            Team {
-                name: "Beijing".into(),
-                score: 27,
-            },
-            Team {
-                name: "Guangzhou".into(),
-                score: 22,
-            },
-            Team {
-                name: "Shandong".into(),
-                score: 12,
-            },
-        ],
-    };
-    b.iter(|| teams.render().unwrap());
+    let teams = Teams::default();
+    b.iter(|| black_box(&teams).render().unwrap());
+}
+
+fn teams_fmt(b: &mut criterion::Bencher) {
+    let teams = Teams::default();
+    b.iter(|| black_box(&teams).to_string());
+}
+
+fn teams_io(b: &mut criterion::Bencher) {
+    let teams = Teams::default();
+    b.iter(|| {
+        let mut vec = Vec::with_capacity(BigTable::SIZE_HINT);
+        black_box(&teams).write_into(&mut vec).unwrap();
+        vec
+    });
 }
 
 #[derive(Template)]
@@ -61,6 +76,32 @@ fn teams(b: &mut criterion::Bencher) {
 struct Teams {
     year: u16,
     teams: Vec<Team>,
+}
+
+impl Default for Teams {
+    fn default() -> Self {
+        Teams {
+            year: 2015,
+            teams: vec![
+                Team {
+                    name: "Jiangsu".into(),
+                    score: 43,
+                },
+                Team {
+                    name: "Beijing".into(),
+                    score: 27,
+                },
+                Team {
+                    name: "Guangzhou".into(),
+                    score: 22,
+                },
+                Team {
+                    name: "Shandong".into(),
+                    score: 12,
+                },
+            ],
+        }
+    }
 }
 
 struct Team {