```text
$ cd testing
$ cargo bench --bench all

Big table (io)          time:   [109.07 µs 109.25 µs 109.45 µs]
                        change: [-42.044% -41.581% -41.182%] (p = 0.00 < 0.05)
                        Performance has improved.

Teams (io)              time:   [265.28 ns 266.07 ns 267.12 ns]
                        change: [-20.221% -19.898% -19.538%] (p = 0.00 < 0.05)
                        Performance has improved.
```