@Kijewski 2024-09-13
Also, I knew that `.to_string()` imposes an overhead, but not that it was this bad:

```text
Big table               time:   [106.57 µs 106.69 µs 106.81 µs]
Big table (fmt)         time:   [189.68 µs 189.85 µs 190.03 µs]
Big table (io)          time:   [104.30 µs 104.39 µs 104.49 µs]

Teams                   time:   [183.76 ns 184.16 ns 184.65 ns]
Teams (fmt)             time:   [407.84 ns 408.43 ns 409.03 ns]
Teams (io)              time:   [261.91 ns 262.31 ns 262.75 ns]
```

We should probably mention it in the book.