@GuillaumeGomez 2024-09-14
Wow, that's a crazy perf improvement. Well done!